requirements.txt
openai
pyttsx3
pyserial
python-dotenv
langdetect